Code comments (#51)
* Add comments.

* Add comments. Fix whitespace

* Add comments. Fix whitespace

* Fix whitespace

* Fix whitespace.

* Format document and fix whitespace.

* Format document.

* Format document and add comments.

* Add _target=blank attr to image credits.

* Add guard statement to check ownership of book
before deletion

* Add cancel button to signout confirmation page

* Change theme

* Add section in lint results.

* Initial commit

* Enhance description of BDD testing.

* Fix whitespace formatting

* Update TOC in README

+++ shelf/views.py
@@ -9,8 +9,6 @@ from django.core.paginator import Paginator, EmptyPage, PageNotAnInteger
 from .models import Book, Comment
 from .forms import BookForm, CommentForm
 
-# Create your views here.
-
 
 def load_home_page(request):
     if request.user.is_authenticated:
@@ -48,6 +46,7 @@ def get_books(request):
         return redirect('accounts/login')
 
 
+# This view displays the form for the user to add a new Book object.
 def add_book_form(request):
     if request.user.is_authenticated:
         book_form = BookForm()
@@ -60,6 +59,7 @@ def add_book_form(request):
         return redirect('accounts/login')
 
 
+# This view adds the book to the database.
 def add_book(request):
     if request.method == 'POST':
 
@@ -72,22 +72,24 @@ def add_book(request):
             slug = slugify(title)
             image_url = Book.COVER_URLS.get(genre)
             Book.objects.create(title=title, author=author, description=description,
-                            genre=genre, owner=request.user, slug=slug, image_url=image_url)
+                                genre=genre, owner=request.user, slug=slug, image_url=image_url)
             messages.info(request, "Successfully added your book.")
             return redirect('/')
         else:
-            messages.error(request, "Woops. Something went wrong. Could not save your book.")
+            messages.error(
+                request, "Woops. Something went wrong. Could not save your book.")
             context = {
                 'form': book_form
             }
             return render(request, 'add_book_template.html', context)
 
 
+# The method opens the BookForm populated with the current data so the user can update it.
 def edit_book(request, book_id):
     book = get_object_or_404(Book, id=book_id)
-    
     if request.method == 'POST':
         form = BookForm(request.POST, instance=book)
+
         if form.is_valid():
             form.save()
             updated_book = get_object_or_404(Book, id=book_id)
@@ -98,6 +100,7 @@ def edit_book(request, book_id):
             return redirect('/')
         else:
             messages.error(request, "Woops. Something went wrong.")
+
     form = BookForm(instance=book)
     context = {
         'form': form
@@ -108,8 +111,11 @@ def edit_book(request, book_id):
 
 def delete_book(request, book_id):
     book = get_object_or_404(Book, id=book_id)
-    book.delete()
-    messages.info(request, "Book succssfully deleted.")
+    if book.owner == request.user:
+        book.delete()
+        messages.info(request, "Book succssfully deleted.")
+    else:
+        messages.error(request, "Could not delete that book.")
     return redirect('/')
 
 
@@ -125,19 +131,18 @@ def add_comment(request, book_id):
     else:
         messages.error(request, "Something went wrong.")
         comment_form = CommentForm()
-    
     # Paginate the comments
     comments = book.comments.order_by("-written_on")
     page = request.GET.get('page', 1)
     paginator = Paginator(comments, 4)
-    
+
     try:
         paginated_comments = paginator.page(page)
     except PageNotAnInteger:
         paginated_comments = paginator.page(1)
     except EmptyPage:
         paginated_comments = paginator.page(paginator.num_pages)
-    
+
     context = {
         'comment_form': CommentForm(),
         'book': book,
@@ -158,11 +163,13 @@ def delete_comment(request, comment_id):
     return redirect(f'/book_detail/{book_id}')
 
 
+# Method to update the comment. It loads the old comment and book to which it belongs,
+# replaces the body of the old comment, and saves it to the book.
 def update_comment(request, comment_id):
     comment = get_object_or_404(Comment, id=comment_id)
     book_id = comment.book.id
 
-    if request.user == comment.author:   
+    if request.user == comment.author:
         new_comment_body = request.POST.get('new-comment-text')
         new_comment_body.strip()
         comment.body = new_comment_body
@@ -182,7 +189,7 @@ def view_book_detail(request, book_id):
     # Paginate the comments
     page = request.GET.get('page', 1)
     paginator = Paginator(comments, 4)
-    
+
     try:
         paginated_comments = paginator.page(page)
     except PageNotAnInteger:
@@ -209,8 +216,9 @@ def remove_like(request, book_id):
     book = get_object_or_404(Book, id=book_id)
     book.likes.remove(request.user)
     return redirect('/')
- 
 
+
+# Method gets all of the books posted by the user.
 def get_my_books(request):
 
     books = Book.objects.filter(owner=request.user)
@@ -233,6 +241,7 @@ def get_my_books(request):
         return redirect('accounts/login')
 
 
+# Method gets all of the books that the user has liked.
 def get_favorites(request):
     books = Book.objects.filter(likes__id=request.user.id)
 
@@ -254,12 +263,13 @@ def get_favorites(request):
         return redirect('accounts/login')
 
 
+# Method gets all of the books belonging to a particular owner.
 def books_by_owner(request, owner_id):
     books = Book.objects.filter(owner__id=owner_id).order_by('-posted_on')
     owner = User.objects.get(id=owner_id)
 
     print(owner.username)
-    
+
     liked_books = []
     for book in books:
         if book.likes.filter(id=request.user.id).exists():
@@ -280,8 +290,12 @@ def books_by_owner(request, owner_id):
         return redirect('accounts/login')
 
 
+# Method uses Q objects to query the database.
 def perform_search(request):
+    # Get the selected genres.
     list = request.POST.getlist('genres')
+
+    # Initialize Q objects to query each category.
     genre_query = Q()
     author_query = Q()
     title_query = Q()
@@ -289,18 +303,22 @@ def perform_search(request):
 
     liked_books = []
 
+    # Add each genre to the Q object as an OR search
     for gen in list:
         genre_query = genre_query | Q(genre__iexact=gen)
-    
+
+    # Get the values from the form fields.
     title_search_terms = request.POST['title-search-input']
     author_search_terms = request.POST['author-search-input']
     description_search_terms = request.POST['description-search-input']
 
+    # If there are terms (ie if the user gave data in the form field)
+    # split the response into an array of strings
+    # add each term to the appropriate Q object as an OR search.
     if title_search_terms:
         terms = title_search_terms.split()
         for term in terms:
             title_query = title_query | Q(title__icontains=term)
-        print(title_query)
 
     if author_search_terms:
         terms = author_search_terms.split()
@@ -310,9 +328,14 @@ def perform_search(request):
     if description_search_terms:
         terms = description_search_terms.split()
         for term in terms:
-            description_query = description_query | Q(description__icontains=term)
-    
-    books = Book.objects.filter(author_query, title_query, genre_query, description_query)
+            description_query = description_query | Q(
+                description__icontains=term)
+
+    # Filter the book objects by submitting each of the query objects.
+    books = Book.objects.filter(
+        author_query, title_query, genre_query, description_query)
+
+    # Determine which of the books the user has liked.
     for book in books:
         if book.likes.filter(id=request.user.id).exists():
             liked_books.append(book.id)
